q_reserve/requirements.txt
# Core Flask and web framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-SocketIO==5.3.6

# Database and ORM
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Vector database for embeddings
pgvector==0.2.4

# Background jobs and caching
celery==5.3.4
redis==5.0.1
python-redis-lock==4.0.0

# AI and ML
openai==1.3.5
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
numpy==1.24.4
scikit-learn==1.3.2

# Email and templates
Jinja2==3.1.2
MarkupSafe==2.1.3

# Image processing
Pillow==10.1.0

# API and HTTP
requests==2.31.0
httpx==0.25.2

# Validation and serialization
marshmallow==3.20.1
email-validator==2.1.0.post1

# Security
bcrypt==4.1.2
cryptography==41.0.8

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Linting and formatting
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Environment and configuration
python-dotenv==1.0.0
click==8.1.7

# Monitoring and logging
sentry-sdk[flask]==1.38.0

# Date and time
python-dateutil==2.8.2
pytz==2023.3

# File handling
python-magic==0.4.27

# Translation
googletrans==4.0.0rc1
deep-translator==1.11.4

# Web scraping and parsing (for knowledge base)
beautifulsoup4==4.12.2
lxml==4.9.3

# Rate limiting and throttling
slowapi==0.1.9

# Task scheduling
APScheduler==3.10.4

# Webhook support
webhooks==0.1.0

# Production server
gunicorn==21.2.0
gevent==23.9.1